+++ Core/Src/ring_buffer.c
@@ -2,7 +2,7 @@
 #include "ring_buffer.h"
 
 
-void ring_buffer_init(ring_buffer_t *rb, uint8_t *mem_add, uint8_t cap)
+void ring_buffer_init(ring_buffer_t *rb, uint16_t *mem_add, uint16_t cap)
 {
 	rb->buffer = mem_add;
 	rb->capacity = cap;
